requirements.txt
requests
opencv-python
numpy
Pillow
